Switch between Zarr and NumPy with one CLI flag (#20)
* Switch between Zarr and NumPy with one CLI flag

Previously, to switch between Zarr and NumPy, users had to make changes
to 2 parts of the config: one for the final transformation, and another
for the sink.
Additionally, the final transformation had to be changed
in code. This was cumbersome.

This PR improves this by using Hydra's interpolation feature, and
setting defaults which will be loaded accordingly based on the CLI
parameter: serialization_format. If that flag is set to be either Zarr
or NumPy, the final transformation will also be changed accordingly.

* Add license header